full raytracer as of book 1

+++ Book 1/camera.h
@@ -8,7 +8,7 @@ vec3 randomInUnitDisk()
     vec3 p;
     do
     {
-        p = 2.0 * vec3(drand48(), drand48(), 0) - vec3(1, 1, 0);
+        p = 2.0 * vec3((float) drand48(), (float) drand48(), 0) - vec3(1, 1, 0);
     } while (dot(p, p) >= 1.0);
     return p;
 }
