+++ Book 1/diffuse.cpp
@@ -12,7 +12,7 @@ vec3 randomInUnitSphere()
     vec3 p;
     do
     {
-        p = 2.0 * vec3(drand48(), drand48(), drand48()) - vec3(1, 1, 1);
+        p = 2.0 * vec3((float) drand48(), (float) drand48(), (float) drand48()) - vec3(1, 1, 1);
     } while (p.squaredLength() >= 1.0);
     return p;
 }
@@ -61,8 +61,8 @@ int main()
             for (int s = 0; s < ns; s++)
             {
 
-                float u = float(i + drand48()) / float(nx);
-                float v = float(j + drand48()) / float(ny);
+                float u = float(i + (float) drand48()) / float(nx);
+                float v = float(j + (float) drand48()) / float(ny);
                 ray r = cam.getRay(u, v);
                 col += color(r, world);
             }
